Move UpstreamReceiver to Hono Messaging.
The UpstreamReceiver is not used by any other service, thus it makes
sense to remove it from service-base.
