+++ services/messaging/src/main/java/org/eclipse/hono/messaging/MessageDiscardingDownstreamAdapter.java
@@ -16,7 +16,6 @@ import java.util.Map;
 import java.util.function.Consumer;
 
 import org.apache.qpid.proton.message.Message;
-import org.eclipse.hono.service.amqp.UpstreamReceiver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Profile;
