+++ services/messaging/src/main/java/org/eclipse/hono/messaging/UpstreamReceiver.java
@@ -10,7 +10,7 @@
  *    Bosch Software Innovations GmbH - initial creation
  */
 
-package org.eclipse.hono.service.amqp;
+package org.eclipse.hono.messaging;
 
 import org.apache.qpid.proton.amqp.transport.ErrorCondition;
