+++ services/messaging/src/test/java/org/eclipse/hono/TestSupport.java
@@ -28,7 +28,7 @@ import org.apache.qpid.proton.message.Message;
 import org.eclipse.hono.auth.HonoUser;
 import org.eclipse.hono.connection.ConnectionFactory;
 import org.eclipse.hono.messaging.SenderFactory;
-import org.eclipse.hono.service.amqp.UpstreamReceiver;
+import org.eclipse.hono.messaging.UpstreamReceiver;
 import org.eclipse.hono.service.auth.AuthenticationService;
 import org.eclipse.hono.util.Constants;
 import org.eclipse.hono.util.MessageHelper;
